requirements.txt
mcunet
torch
torchvision
Pillow>=9.0.0
psutil>=5.9.0
